Extract axis rendering and prepare shared memory ring buffers
Introduced in this commit is the abstraction of the rendering classes into the `MSRenderer` protocol, each of which provides a customized drawing function under the protocol method stub. More importantly, support has been added (though incompletely) to allow a ring buffer to share memory instead of creating multiple MTLBuffer instances. The alignment constraints are not working as expected with the constant address space, which has stricter alignment rules on macOS, but this will be patched in the next commit.

A bug in the particle renderer was also patched. The MSTileBuffer holding memory to the two particle buffers continued to switch read and write assignments while the simulation was paused, causing the particles to jiggle between the two different positions stored in the two buffers. Additionally, it was discovered that within the ThreadgroupParticleKernel conversions were made from ushort to uint implicitly. This commit makes that conversion explicit

+++ MetalSpheres/Rendering and Physics/Metal Files/ParticlePhysics.metal
@@ -12,6 +12,7 @@
 
 // Metal Files
 #import <metal_stdlib>
+#import <metal_matrix>
 #import "PhysicsCompute.metal"
 #import "Uniforms.metal"
 #import "MetalExtension.metal"
@@ -65,11 +66,11 @@ kernel void ThreadgroupParticleKernel(constant uint &maximumValidThreadIndex    
     if (maximumThreadIndex == 0 || !(simulateGravity || simulateElectrostatics)) return;
         
     // The index in the thread dispatch
-    const uint gridIndex { static_cast<ushort>(threadPos.x + threadsPerGrid.x * threadPos.y) };
+    const uint gridIndex { static_cast<uint>(threadPos.x + threadsPerGrid.x * threadPos.y) };
     
     // Index into the shared threadgroup memory for this particular thread
-    const uint threadgroupIndex { static_cast<ushort>(threadgroupPos.x + threadsPerThreadgroup.x * threadgroupPos.y) };
-    const uint threadsInThreadgroup { static_cast<ushort>(threadsPerThreadgroup.x * threadsPerThreadgroup.y) };
+    const uint threadgroupIndex { static_cast<uint>(threadgroupPos.x + threadsPerThreadgroup.x * threadgroupPos.y) };
+    const uint threadsInThreadgroup { static_cast<uint>(threadsPerThreadgroup.x * threadsPerThreadgroup.y) };
     
     // Represents the current index in the threadgroup seen by this thread as it walks through the memory
     thread uint i = 0;
